Set up text matrix for e2e testing on different operating systems
Update the "Check / Test end-to-end" job to run on MacOS and Windows in
addition to Ubuntu.

Like the runner images for other jobs, the MacOS and Windows images are
pinned to a specific version to promote reproducibility. To avoid the
name of the job changing when the runner image is updated, an explicit
name is used.