Support OSX (#2)
* Set up text matrix for e2e testing on Ubuntu and OSX

Update the "Check / Test end-to-end" job to run on OSX in addition to
Ubuntu.

Like the runner images for other jobs, the OSX image is pinned to a
specific version to promote reproducibility. To avoid the name of the
job changing when the runner image is updated, an explicit name is used.

* Fix `mkdir` invocation for OSX

The standard `mkdir` binary on OSX does not appear to support the full
version `--parents`, so we'll use `-p` instead.

* Fix `cp` invocation for OSX

The standard `cp` binary on OSX does not appear to support the full
version `--recursive`, so we'll use `-r` instead.

* Update the changelog

Include change entry for the changes to `mkdir` and `cp` invocation that
fixes errors for OSX when trying to download or install yamllint.