Fallback to `python` if `python3` is not available
Add a fallback for using the `python` executable if `python3` is not
available on the system. Only use it if `python` is a Python v3
executable, because if it is not it might be even more confusing (given
that yamllint requires Python v3).

Implementation is based on the `sha256sum` fallback logic for `shasum`.