Add and use `shfmt` (#12)
Add shfmt [1] as a tool and use it in two new Make targets called
"format" and "format-check". format runs shfmt in --write mode, updating
the formatting of shell scripts in place. format-check runs shfmt in
--diff mode, checking if shell scripts are formatted and exiting with a
non-zero exit code if not.

A new continuous integration check related to formatting has been added
to the `check.yml` workflow. This simply checks if shell scripts are
formatted with shfmt.

The formatting is applied to all shell scripts in the project, this only
affected `lib/common.sh`.

Lastly, as a tool included in `.tool-versions`, shfmt has also been
added to the `nightly.yml` tooling update workflow.

--
1. https://github.com/mvdan/sh