Avoid repeating target names in the Makefile
Update the `test-download` and `test-install` Make targets to be defined
only once (and condition only the implementation rather than target and
implementation).

This simplifies maintenance, due to the removal of duplicated names, at
the cost of maybe sometimes unnecessarily creating the temporary dir. In
particular, the temp dir will now be created for these targets even if
it won't be used. This is justified because, in all likelihood, after
invoking these targets incorrectly the contributor will invoke it
correctly and needing the temp dir anyway.