Automatically run `asdf install`
Automate running `asdf install` for targets that use tooling defined in
the .tool-versions file. This is achieved by creating a temporary file
that needs to be updated if .tool-versions has changed, and setting this
file as a prerequisite for targets using such tooling.

Since this project is a plugin for asdf, it is considered reasonable to
assume that contributors have asdf installed and available on their
system.

Accordingly, asdf is now listed as a proper prerequisite in the
Contributing Guidelines accordingly. (Also, sneak in the addition of
shfmt [1] to the Contributing Guidelines.)

--
1. 99871d2029802e8ca1dbad68e3fb086cb2e8d936