Improve the `Makefile` (#14)
* Avoid repeating target names in the Makefile

Update the `test-download` and `test-install` Make targets to be defined
only once (and condition only the implementation rather than target and
implementation).

This simplifies maintenance, due to the removal of duplicated names, at
the cost of maybe sometimes unnecessarily creating the temporary dir. In
particular, the temp dir will now be created for these targets even if
it won't be used. This is justified because, in all likelihood, after
invoking these targets incorrectly the contributor will invoke it
correctly and needing the temp dir anyway.

* Create convenience targets `lint` and `verify`

Create two new Make targets for the contributors convenience, avoiding
having to (remember and) invoke multiple commands. First, `lint` can
be used to run all `lint-*` targets. Second, `verify` can be used to
run all checks that should (~)always succeed.

The `.PHONY` target has been updated to include these two new targets
(interestingly, `lint` was already included, oops...)

* Rewrite `.PHONY` target

Update the `.PHONY` target for readability. Instead of one long line
that get's difficult to read (and may be hard to diff), phony targets
are now listed on separate lines grouped roughly by clear categories.

* Automate script detection for formatting and linting

Use wildcard expansion to automatically find `ALL_SCRIPTS`. Use this
variable for formatting and linting.

* Avoid repeating the name of the (top-level) bin directory

* Automatically run `asdf install`

Automate running `asdf install` for targets that use tooling defined in
the .tool-versions file. This is achieved by creating a temporary file
that needs to be updated if .tool-versions has changed, and setting this
file as a prerequisite for targets using such tooling.

Since this project is a plugin for asdf, it is considered reasonable to
assume that contributors have asdf installed and available on their
system.

Accordingly, asdf is now listed as a proper prerequisite in the
Contributing Guidelines accordingly. (Also, sneak in the addition of
shfmt from 99871d2 to the Contributing Guidelines.)