Remove prerequisites (#18)
Remove all prerequisites checks. They're of limited value (scripts will
fail due to them missing anyway), hard to maintain (manual review is
required to see if prerequisites are still up-to-date, this can be
especially problematic when a prerequisite is no longer necessary), and
can't cover more complex cases such as the use of `python3` or `python`.

Include this change in the changelog as well.