Define development environment with Docker
Create a Dockerfile that can be used to instantiate a container for
development purposes. It is configured to install the necessary
prerequisites (both for asdf-vm and this project), and install+setup
asdf.

Due to an uncooperative version of the `sha256sum` binary in the
container the 'test-download', 'test-install', and 'test-installation'
targets currently don't work inside the container.

Also, define a Make command for running using the Dockerfile to run an
ephemeral development environment. This command will build the image (if
necessary) and run it mounted to the current working directory.

Lastly, configure Dependabot to keep the Dockerfile base image
up-to-date.