Set up hadolint to lint the project's Dockerfile
Add hadolint [1] as a development dependency to be used to lint the
Dockerfile that was added as a development environment definition. It's
version is managed through asdf (see `.tool-versions`) and it's
configured through the `.hadolint.yml` file.

The hadolint is configured to ignore two rules. First DL3018 [2] as
there's currently no need to pin the versions of dependencies installed
with `apk`. Second SC1091 [3] to avoid warnings about sourcing asdf.

hadolint will be run as part of `make lint` and the "Check/Lint" CI job.

--
1. https://github.com/hadolint/hadolint
2. https://github.com/hadolint/hadolint/wiki/DL3018
3. https://www.shellcheck.net/wiki/SC1091