Continuously validate development image
Set up a CI job, "Check / Dev env", to check the validity of the dev env
image. For now, it just verifies the image can be build, this may be
expanded in the future.

To accommodate this the Make target 'dev-img' was created. This provides
a more permanent API than the file '.tmp/.dev-img' for building the dev
image. Additionally, it provides a way for contributors to build the dev
image and run it in their own way.