Define development environment (#22)
Create a Dockerfile [1] that can be used to instantiate a container for
development purposes. It is configured to install the necessary
prerequisites (both for asdf-vm and this project), and install+setup
asdf.

Also define a Make target - `dev-env` - for running using the Dockerfile
to run an ephemeral development environment. This command will build the
image (if necessary) and run it mounted to the current working directory

It is also possible to build the image without running it with the
target `dev-img`. This is used in CI to validate the dev environment can
be build continuously and allows contributors to run the dev environment
in their preferred way.

Add hadolint [2] as a development dependency to be used to lint the
Dockerfile It's version is managed through asdf (see `.tool-versions`)
and it's configured through the `.hadolint.yml` file.

The hadolint is configured to ignore two rules. First DL3018 [3] as
there's currently no need to pin the versions of dependencies installed
with `apk`. Second SC1091 [4] to avoid warnings about sourcing asdf.

hadolint is run as part of `make lint` and the "Check / Lint" CI job.

Lastly, configure Dependabot to keep the Dockerfile base image
up-to-date.

--
1. https://www.docker.com/
2. https://github.com/hadolint/hadolint
3. https://github.com/hadolint/hadolint/wiki/DL3018
4. https://www.shellcheck.net/wiki/SC1091