Find download URL dynamically
Update the logic that determines the download URL for the yamllint
tarball to search for the `urls` entry with the correct(?) package type
instead of just using the second entry. This should help ensure the
selected URL is always for a tarball. This is necessary as various
(most?) yamllint releases also include a URL for a `.whl` file.

(?): I did not find official documentation on the structure of the JSON
returned by PyPI, so the condition is instead based on the semantics of
observed JSON and testing of the changes on various yamllint releases.
If this turns out to be incorrect/insufficient, a patch may be applied
later.