Avoid error due to directives in requires.txt
Update the yamllint installation logic to **avoid** an installation
error first observed with yamllint v1.31.0 due to a `[dev]` directive
in the egg-info requires.txt file. The use of the word _avoid_ is
intentional as this is probably not the best/correct solution - at least
it works as a temporary solution (at the cost of slower installation
times due to having to install development dependencies?).

Also updated the `make test-installation` target to include the
outputting the installed version. This should help during development to
be able to tell if the nth (n>1) installation actually worked, or you're
just testing the last successful installation.