Document prerequisite tooling in the README
Add a new section to the README with the prerequisite tooling to be able
to use this asdf plugin.

This follows earlier attempts, e.g. [1], to check for these at runtime.
This became unwieldy and was removed in [2]. The approach taken here is
inspired by _the_ asdf-syft plugin [3] which has a "Dependencies"
section in its documentation.

--
1. 2307d6e3dbc7d9d6d2f729c64b71333c06f9ce2d
2. 174bb5bd90e05e376734c3584e6f8293230b7ae4
3. https://github.com/davidgp1701/asdf-syft