Improve nightly tooling update workflow (#58)
* Rename secrets for tooling update for clarity

The previous names were copied from another project and the names don't
make sense in the context in which they're used here. Since these
values
can be used for various automation purposes, this current names were
chosen.

* Use `ericcornelissen/tool-versions-update-action` for tooling updates

Replace the home-grown tooling update logic for this action I build for
the purpose of updating tooling. The action will work better in the long
run as improvements to it become available.