Add all (relevant) help scripts
Create the help scripts for the plugin as prescribed in [1]. All help
scripts have been created except for `bin/help.config`, which doesn't
apply to this plugin (no configuration is requried, as far as I know).

For convenience, a Make target that runs each help script (in some
sensible order) has been added for a quick look at the full help
message. For now, when working on a single help script, it can be
invoked directly for testing purposes.

--
1. https://asdf-vm.com/plugins/create.html