Update `Containerfile` to not use Docker-specific `SHELL` directive
Remove the `SHELL` directive from the `Containerfile` because it is
specific to Docker. Just removing it results in build errors in the
second `RUN` directive, to avoid this the asdf installation steps have
been changed to match the "POSIX Shell & Git" installation instructions
so that it's installed in a way compatible with any shell. Since in the
end the container entrypoint is still Bash, the source instruction is
still inserted into the `.bashrc` configuration file.