Replace `Dockerfile` with `Containerfile` (#83)
Replace the `Dockerfile` with a `Containerfile`. Update it to not use
Docker-specific `SHELL` directive. Just removing it results in build
errors in the second `RUN` directive. To avoid this the asdf install
steps have been changed to match the "POSIX Shell & Git" installation
instructions of asdf so that it's installed in a way compatible with any
POSIX compliant shell. (Since in the end the container entrypoint is
still Bash, the source instruction is still inserted into the `.bashrc`
configuration file.)

Dependabot does not currently support applying updates to Containerfile,
so the `docker` option is removed from the configuration. When it is
supported, it would ideally be reconfigured if Dependabot is still used.