Update yamllint installation to utilize a Python venv
Update the `install_version` script to create a venv when installing
yamllint. The venv is local to the yamllint version and is active only
during installation and when running yamllint through the helper script
that is available on the path as yamllint.

Additionally, this does some minor refactoring on the `install_version`
script to reduce duplication.