Use `--quiet` when using `sha256sum`
This partially reverts deba6987c5e1cc1adada245848c708a0f5b84fc2 by again
using a non-shorthand flag when calling `sha256sum`. This follows a bug
report from a MacOS Sonoma 14.4 user stating that `-s` isn't supported.
I verified this on an Ubuntu 22.04 machine where I observed the same.
Unfortunately the reference commit does not contain enough information
about whether or not this would have worked back then. For now, I'm
inclined to go with the non-shorthand version because, at least for
newer systems, it seems to be the best option. Hopefully this suffices,
if not we may have to find some more advanced solution that not only
checks which command is available but also what flags (e.g. based on
the `--help` message).

This commit reverts the previous two commits on this branch as they were
intended as temporary commits to be able to run tests in CI.