Dynamically choose flags for `sha256sum`
When using `sha256sum` to validate checksums, in order to avoid
outputting unnecessary information we want to use the `--quiet` flag.
However, some implementations don't support `--quiet` and instead
support `-s` (more on this below). To support this we now dynamically
check which is available by looking for the string "--quiet" in the help
message of the `sha256sum` program. This is achieved using `grep`, so
that is now included in the list of dependencies for this plugin.

From testing, it seems `-s` is supported by the Busybox implementation
of `sha256sum` (as seen in the ephemeral development image of this
project, to see this run `make dev-env` and then `sha256sum --help` on
this commit). Other implementation, based on the cause of this change
those shipped with Ubuntu 22.04 and MacOS Sonoma 14.4, only support
`--quiet`.