Rely on redirection rather than CLI flags to suppress output
Update the use of `shasum`/`sha256sum` to omit the use of flags to
suppress output (`-q`, `-s`, or `--quiet`) and instead redirect stdout
to `/dev/null` in order to suppress it. This avoids complications due to
different implementations of the same program and _should_ be fine given
this project is only expected to be used in Unix environments.