Fix `sha256sum: invalid option` (#127)
This partially reverts deba6987c5e1cc1adada245848c708a0f5b84fc2 by
changing the use of flags for `shasum` and `sha256sum`. This follows a
bug report from a MacOS Sonoma 14.4 user stating that `-s` isn't
supported. I verified this on an Ubuntu 22.04 machine where I observed
the same. Unfortunately the reference commit does not contain the info
necessary to determine on which system it didn't work back then, but the
ephemeral development environment for this project (`make dev-env`) has
a Busybox implementation which does support `-s` (but not `--quiet`,
which appears supported on MacOS and Ubuntu).

The chosen fix is to rely on redirection rather than CLI flags to
suppress the output. To achieve that, this updates the use of `shasum`/
`sha256sum` to omit the use of flags to suppress output (`-q`, `-s`, or
`--quiet`) and instead redirect stdout to `/dev/null` in order to
suppress it. This avoids complications due to different implementations
of the same program and _should_ be fine given this project is only
expected to be used in Unix environments.


An alternative that was considered is to dynamically choose flags for
`sha256sum`. To support this we could dynamically check which is
available by looking for the string "--quiet" in the help message of the
`sha256sum` program. This could be achieved using `grep` using something
like:

    if [ "$(sha256sum --help 2>&1 | grep -c -- --quiet)" -gt 0 ]; then
      shasum_command='sha256sum --quiet'
    else
      shasum_command='sha256sum -s'
    fi

(This would require adding "grep" to `bin/help.deps` as it's a new
project dependency.)