Create explicit uninstall script
Create the optional `bin/uninstall` script based on the docs[1], the
asdf-erlang uninstall script (at 826ad0e), and manual debugging. In
essence, the script just removes the directory where the version is
installed, because of the use of venv everything related to yamllint
is present in that directory.

A `test-` make target for uninstall was added too, to support manually
testing the functionality. This test is also executed in CI.

--
1. https://asdf-vm.com/plugins/create.html#bin-uninstall