Remove secrets check from CI
Update the CI to omit the job that checks for secrets and instead rely solely on GitHub for secret scanning in a continuous fashion. Individual contributors can still continue to use any secret scanning software they like locally.

The motivation for this removal is to reduce the supply chain surface by omitting unnecessary dependencies. This dependencies adds limited value (given what GitHub already does for us), requires access to secrets, and has a history of unreliability.