Create explicit `latest-stable` script
Create the optional `bin/latest-stable` script based on the docs[1]
description of what it should do. It will use the PyPI API to determine
the latest stable version.

A `test-` make target for the script was added too, to support manually
testing the functionality. This test is also executed in CI.

--
1. https://asdf-vm.com/plugins/create.html#bin-latest-stable