Improve license holders (#150)
be explicit about who the license holders are since "asdf-yamllint
Contributors" isn't a meaningful term to refer to. Using individual
names is more explicit and "safer". The names are based on the displayed
name for each contributor on GitHub at the time of this commit.

See also 3f2035de7afbbb3d319bd27d6635c90466416c8b