Remove incorrect `set -eo pipefail`
Remote the directive from the sourced helper script which should not be
setting it. In the case of this project it appears harmless because all
the main scripts set it anyway, but it could have negative consequences.
As I understand it, scripts that are `source`d like this should not set
it because it could affect the behavior of the sourcing script
unexpectedly.