Replace Semgrep PRO with Semgrep OSS
Remove `semgrep.yml`, switching from Semgrep PRO to Semgrep OSS, and
running semgrep as part of the `check.yml` workflow. The tradeoff here
is giving up some SAST rules and SSC support (which is n/a anyway) in
favor of not needing an authentication token, running for all changes,
allowing anyone to perform the scan locally, and no dependence on the
availability of the Semgrep backend.