Implement explicit error handling for helpers (#151)
Remote the directive from the sourced helper script which should not be
setting it. In the case of this project it appears harmless because all
the main scripts set it anyway, but it could have negative consequences.
As I understand it, scripts that are `source`d like this should not set
it because it could affect the behavior of the sourcing script
unexpectedly. Correspondingly, update all common functions to include
proper error handling.

Also, simplify usage of the `check_env_var` by using indirect environment
variable lookup (rather than passing the value explicitly).