Update the security policy (#176)
* Reword the security policy in terms of latest release only

Update the security policy so that only the latest release will be
supported going forward. This change is primarily because I don't really
expect there to ever be a second major version, and if there is we can
decide at that time whether it makes sense to support both major
versions.

* Bump intended security support until 2026-01-01